requirements.txt
Flask==3.1.1
mysql-connector-python==8.0.26
